Hardcoded kickoffs, made it look as silly as it deserves too
