+++ RLBotPack/TensorBot/appearance.cfg
@@ -1,0 +1,100 @@
+[Bot Loadout]
+# Primary Color selection
+team_color_id = 9
+# Secondary Color selection
+custom_color_id = 0
+# Car type (Octane, Merc, etc)
+car_id = 607
+# Type of decal
+decal_id = 4989
+# Wheel selection
+wheels_id = 4795
+# Boost selection
+boost_id = 5112
+# Antenna Selection
+antenna_id = 0
+# Hat Selection
+hat_id = 3134
+# Paint Type (for first color)
+paint_finish_id = 1681
+# Paint Type (for secondary color)
+custom_finish_id = 1681
+# Engine Audio Selection
+engine_audio_id = 0
+# Car trail Selection
+trails_id = 3220
+# Goal Explosion Selection
+goal_explosion_id = 3071
+# Finds the closest primary color swatch based on the provided RGB value like [34, 255, 60]
+primary_color_lookup = None
+# Finds the closest secondary color swatch based on the provided RGB value like [34, 255, 60]
+secondary_color_lookup = None
+
+[Bot Loadout Orange]
+# Primary Color selection
+team_color_id = 19
+# Secondary Color selection
+custom_color_id = 0
+# Car type (Octane, Merc, etc)
+car_id = 607
+# Type of decal
+decal_id = 4989
+# Wheel selection
+wheels_id = 4795
+# Boost selection
+boost_id = 5112
+# Antenna Selection
+antenna_id = 0
+# Hat Selection
+hat_id = 3134
+# Paint Type (for first color)
+paint_finish_id = 1681
+# Paint Type (for secondary color)
+custom_finish_id = 1681
+# Engine Audio Selection
+engine_audio_id = 0
+# Car trail Selection
+trails_id = 3220
+# Goal Explosion Selection
+goal_explosion_id = 3071
+# Finds the closest primary color swatch based on the provided RGB value like [34, 255, 60]
+primary_color_lookup = None
+# Finds the closest secondary color swatch based on the provided RGB value like [34, 255, 60]
+secondary_color_lookup = None
+
+[Bot Paint Blue]
+# car_paint_id
+car_paint_id = 12
+# decal_paint_id
+decal_paint_id = 0
+# wheels_paint_id
+wheels_paint_id = 8
+# boost_paint_id
+boost_paint_id = 8
+# antenna_paint_id
+antenna_paint_id = 0
+# hat_paint_id
+hat_paint_id = 0
+# trails_paint_id
+trails_paint_id = 8
+# goal_explosion_paint_id
+goal_explosion_paint_id = 0
+
+[Bot Paint Orange]
+# car_paint_id
+car_paint_id = 12
+# decal_paint_id
+decal_paint_id = 0
+# wheels_paint_id
+wheels_paint_id = 8
+# boost_paint_id
+boost_paint_id = 8
+# antenna_paint_id
+antenna_paint_id = 0
+# hat_paint_id
+hat_paint_id = 0
+# trails_paint_id
+trails_paint_id = 8
+# goal_explosion_paint_id
+goal_explosion_paint_id = 0
+
