+++ RLBotPack/TensorBot/bot.cfg
@@ -23,7 +23,7 @@ requirements_file = ./requirements.txt
 developer = Soren
 
 # Short description of the bot
-description = 100% Machine Learning bot trained with recorded Diablo data
+description = Machine Learning bot trained with recorded Diablo data
 
 # Fun fact about the bot
 fun_fact = TensorBot is trained on ~60 hours of Diablo recordings
