--- RLBotPack/TensorBot/src/appearance.cfg
@@ -1,49 +1,0 @@
-[Bot Loadout]
-team_color_id = 60
-custom_color_id = 0
-car_id = 26
-decal_id = 0
-wheels_id = 1565
-boost_id = 35
-antenna_id = 0
-hat_id = 0
-paint_finish_id = 1681
-custom_finish_id = 1681
-engine_audio_id = 0
-trails_id = 3220
-goal_explosion_id = 3018
-
-[Bot Loadout Orange]
-team_color_id = 3
-custom_color_id = 0
-car_id = 26
-decal_id = 0
-wheels_id = 1565
-boost_id = 35
-antenna_id = 0
-hat_id = 0
-paint_finish_id = 1681
-custom_finish_id = 1681
-engine_audio_id = 0
-trails_id = 3220
-goal_explosion_id = 3018
-
-[Bot Paint Blue]
-car_paint_id = 12
-decal_paint_id = 0
-wheels_paint_id = 7	
-boost_paint_id = 7
-antenna_paint_id = 0
-hat_paint_id = 0
-trails_paint_id = 2
-goal_explosion_paint_id = 0
-
-[Bot Paint Orange]
-car_paint_id = 12
-decal_paint_id = 0
-wheels_paint_id = 14
-boost_paint_id = 14
-antenna_paint_id = 0
-hat_paint_id = 0
-trails_paint_id = 14
-goal_explosion_paint_id = 0
